Update README.md
In today world we can see lots of opportunity is there for us to get any groceries or vegetables
within 10 minutes to 30 minutes in city areas but we can not get medicine within 10 to 30 min
through any app or delivery partner. So, we are here to solve this problem. We are trying to make a
pharmacy shop chain for 10 to 30 min fast supply of medicines. The main uniqueness of this
project is the one hour delivery of medicines and a quick phone call or chat service with a qualified
doctors. In every city small to large we will connect all medicine shops in one app and they all can
sell their medicines through our app. The cost of medicines will be decided as per dealer shop
buying price in that particular place. We already have made a website for this project. And app
making is on the way for the project. We also will have an emergency ambulance service for
nearest hospital.
Customer service and customer satisfaction is our main goal. Be safe be happy with medicox